// frontend/src/api/models/PlotJSON.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PlotJSON = {
    /**
     * The data for the plot
     */
    data: Array<Record<string, any>>;
    /**
     * The layout for the plot
     */
    layout: Record<string, any>;
};